+++ aecmk/localcert/keyprovider_prego117_windows.go
@@ -1,0 +1,14 @@
+//go:build !go1.17
+// +build !go1.17
+
+package localcert
+
+import (
+	"crypto/x509"
+	"fmt"
+)
+
+func (p *LocalCertProvider) loadWindowsCertStoreCertificate(path string) (privateKey interface{}, cert *x509.Certificate) {
+	panic(fmt.Errorf("Windows cert store not supported until Go 1.17"))
+	return
+}
